refactor: use platform specific file extensions instead
