+++ example/src/utils/integrity/integrity.ios.ts
@@ -1,20 +1,33 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import {
   generateHardwareKey,
   isAttestationServiceAvailable,
 } from "@pagopa/io-react-native-integrity";
+import type { HardwareSignatureWithAuthData } from "src/utils/integrity";
 
 /**
  * Generates the hardware backed key for iOS.
  * It differents from Android because the key is generated via `io-react-native-integrity` instead of `io-react-native-crypto`.
  * @returns a promise that resolves with the key tag as string or rejects with an error.
  */
-const generateIntegrityKeyIos = async () => await generateHardwareKey();
+const generateIntegrityHarwareKeyTag = async () => await generateHardwareKey();
 
 /**
  * Ensures that the integrity servicy is availabe on the device.
  * @returns a promise with resolves with a boolean value indicating wheter or not the integrity service is available.
  */
-const ensureIntegrityServiceIos = async () =>
+const ensureIntegrityServiceIsReady = async () =>
   await isAttestationServiceAvailable();
 
-export { generateIntegrityKeyIos, ensureIntegrityServiceIos };
+const getHardwareSignatureWithAuthData = async (
+  hardwareKeyTag: string,
+  clientData: string
+): Promise<HardwareSignatureWithAuthData> => {
+  return Promise.resolve({ signature: "", authenticatorData: "" });
+};
+
+export {
+  generateIntegrityHarwareKeyTag,
+  ensureIntegrityServiceIsReady,
+  getHardwareSignatureWithAuthData,
+};
