+++ tsconfig.json
@@ -23,6 +23,7 @@
     "skipLibCheck": true,
     "strict": true,
     "target": "esnext",
-    "verbatimModuleSyntax": true
+    "verbatimModuleSyntax": true,
+    "moduleSuffixes": [".ios", ".android", ".native", ""]
   }
 }
